transforms: move allocation out of snax-to-func
